+++ typescript/infra/config/environments/testnet4/agent.ts
@@ -356,7 +356,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '09ba565-20251111-105255',
+      tag: '4d35beb-20251117-142647',
     },
     blacklist: [...releaseCandidateHelloworldMatchingList, ...relayBlacklist],
     gasPaymentEnforcement,
@@ -401,7 +401,7 @@ const releaseCandidate: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '09ba565-20251111-105255',
+      tag: '6c48aa6-20251114-115236',
     },
     blacklist: relayBlacklist,
     gasPaymentEnforcement,