+++ typescript/infra/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @hyperlane-xyz/infra
 
+## 19.9.0
+
+### Patch Changes
+
+- Updated dependencies [8c027d852]
+- Updated dependencies [8c027d852]
+  - @hyperlane-xyz/utils@19.9.0
+  - @hyperlane-xyz/sdk@19.9.0
+  - @hyperlane-xyz/helloworld@19.9.0
+
+## 19.8.0
+
+### Patch Changes
+
+- Updated dependencies [500d81246]
+- Updated dependencies [4614a503e]
+- Updated dependencies [2ed21c97d]
+- Updated dependencies [78ff6cd47]
+- Updated dependencies [00b014a3e]
+- Updated dependencies [3f75ad86d]
+  - @hyperlane-xyz/sdk@19.8.0
+  - @hyperlane-xyz/utils@19.8.0
+  - @hyperlane-xyz/helloworld@19.8.0
+
 ## 19.7.0
 
 ### Patch Changes